pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "worker"
version = "2.0.0-beta1"
authors = [
    { name="Mario Winkler", email="mario.winkler@dlr.de" },
    { name="Jonas Eberle", email="jonas.eberle@dlr.de" }
]
description = "A Flowable External Worker implementation with FastAPI acting as worker for Registration Harvester component"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "fastapi[standard]",
    "stactools-landsat",
    "setuptools",
    "registration-library @ git+https://github.com/EOEPCA/registration-library",
    "flowable.external-worker-client @ git+https://github.com/EOEPCA/eoepca-flowable-external-client-python",
    "eodm @ git+https://github.com/geopython/eodm"
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8"
]
test = [
    "pytest"
]

[project.urls]
Homepage = "https://github.com/EOEPCA/registration-harvester"
Issues = "https://github.com/EOEPCA/registration-harvester/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-q --import-mode=importlib --no-header -rfE --disable-warnings --log-level=INFO"
testpaths = [
    "test",
    "integration",
]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"